requirements/main.txt
click
jsonschema
pyyaml